+++ src/main/java/wooteco/subway/config/WebMvcConfig.java
@@ -3,11 +3,12 @@ package wooteco.subway.config;
 import java.util.List;
 
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import wooteco.subway.web.member.LoginMemberMethodArgumentResolver;
-import wooteco.subway.web.member.interceptor.BearerAuthInterceptor;
+import wooteco.subway.web.restcontroller.member.interceptor.BearerAuthInterceptor;
+import wooteco.subway.web.restcontroller.member.methodargumentresolver.LoginMemberMethodArgumentResolver;
 
 @Configuration
 public class WebMvcConfig implements WebMvcConfigurer {
@@ -28,7 +29,7 @@ public class WebMvcConfig implements WebMvcConfigurer {
     }
 
     @Override
-    public void addArgumentResolvers(List argumentResolvers) {
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
         argumentResolvers.add(loginMemberArgumentResolver);
     }
 }
