+++ src/main/java/wooteco/subway/service/favorite/dto/FavoriteRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.service.favorite;
+package wooteco.subway.service.favorite.dto;
 
 import wooteco.subway.domain.favorite.Favorite;
