+++ src/main/java/wooteco/subway/web/restcontroller/member/MemberController.java
@@ -1,4 +1,4 @@
-package wooteco.subway.web.member;
+package wooteco.subway.web.restcontroller.member;
 
 import java.net.URI;
