+++ src/main/java/wooteco/subway/web/restcontroller/member/methodargumentresolver/LoginMemberId.java
@@ -1,4 +1,4 @@
-package wooteco.subway.web.member;
+package wooteco.subway.web.restcontroller.member.methodargumentresolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
