+++ src/main/java/wooteco/subway/web/restcontroller/path/PathController.java
@@ -1,4 +1,4 @@
-package wooteco.subway.web;
+package wooteco.subway.web.restcontroller.path;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
