+++ src/test/java/wooteco/subway/acceptance/member/MyInfoAcceptanceTest.java
@@ -1,7 +1,7 @@
 package wooteco.subway.acceptance.member;
 
 import static org.assertj.core.api.Assertions.*;
-import static wooteco.subway.AuthAcceptanceTest.*;
+import static wooteco.subway.acceptance.AuthAcceptanceTest.*;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
-import wooteco.subway.AcceptanceTest;
+import wooteco.subway.acceptance.AcceptanceTest;
 import wooteco.subway.service.member.dto.MemberResponse;
 import wooteco.subway.service.member.dto.TokenResponse;
