+++ src/test/java/wooteco/subway/acceptance/station/StationAcceptanceTest.java
@@ -7,7 +7,7 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import wooteco.subway.AcceptanceTest;
+import wooteco.subway.acceptance.AcceptanceTest;
 import wooteco.subway.service.station.dto.StationResponse;
 
 public class StationAcceptanceTest extends AcceptanceTest {
