+++ src/test/java/wooteco/subway/service/favorite/FavoriteServiceTest.java
@@ -16,6 +16,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import wooteco.subway.domain.favorite.Favorite;
 import wooteco.subway.domain.favorite.FavoriteRepository;
 import wooteco.subway.domain.station.StationRepository;
+import wooteco.subway.service.favorite.dto.FavoriteRequest;
+import wooteco.subway.service.favorite.dto.FavoriteResponse;
 
 @ExtendWith(MockitoExtension.class)
 public class FavoriteServiceTest {
