+++ src/test/java/wooteco/subway/web/restcontroller/favorite/FavoriteControllerTest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.web.favorite;
+package wooteco.subway.web.restcontroller.favorite;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.BDDMockito.*;
@@ -31,11 +31,11 @@ import org.springframework.web.filter.ShallowEtagHeaderFilter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import wooteco.subway.doc.FavoriteDocumentation;
 import wooteco.subway.infra.JwtTokenProvider;
-import wooteco.subway.service.favorite.FavoriteRequest;
-import wooteco.subway.service.favorite.FavoriteResponse;
 import wooteco.subway.service.favorite.FavoriteService;
+import wooteco.subway.service.favorite.dto.FavoriteRequest;
+import wooteco.subway.service.favorite.dto.FavoriteResponse;
 import wooteco.subway.service.station.dto.StationResponse;
-import wooteco.subway.web.member.LoginMemberMethodArgumentResolver;
+import wooteco.subway.web.restcontroller.member.methodargumentresolver.LoginMemberMethodArgumentResolver;
 
 @ExtendWith(RestDocumentationExtension.class)
 @SpringBootTest
